Split types, make it easier to write games
